python/requirements.txt
Flask
schedule
pymysql
mysql-connector-python
paramiko
netmiko
requests